+++ fandogh_cli/service_commands.py
@@ -1,7 +1,7 @@
 import click
 from .fandogh_client import *
 from .config import get_project_config
-from .presenter import present
+from fandogh_cli.fandogh_presenter.table_presenter import present
 from .utils import format_text, TextStyle, read_manifest
 from .base_commands import FandoghCommand
 from time import sleep
